+++ pddl/fm.h
@@ -305,6 +305,20 @@ void pddlFmRebuild(pddl_fm_t **c,
                    void *userdata);
 
 /**
+ * Sort elements with dis/con-junction.
+ */
+void pddlFmJuncSort(pddl_fm_junc_t *fm,
+                    int (*cmp)(const pddl_fm_t *fm1,
+                               const pddl_fm_t *fm2,
+                               void *userdata),
+                    void *userdata);
+
+/**
+ * Returns number of atoms appearing in the formula.
+ */
+int pddlFmNumAtoms(const pddl_fm_t *fm);
+
+/**
  * When first (when ...) node, that has non-static preconditions, is found,
  * it is removed and returned.
  * If no (when ...) is found, NULL is returned.